+++ RDACode/Hardware/CRC16/CRC16.h
@@ -3,7 +3,7 @@
 
 /* Includes ------------------------------------------------------------------*/
 #include "sys.h"
-uint8_t CRC16(uint8_t *puchMsg, uint16_t usDataLen);
+uint16_t CRC16(uint8_t *puchMsg, uint16_t usDataLen);
 
 
 #endif /* __CRC16_H */
